#426 provided a check if the file being uploaded contains the required objects for database insertion

+++ app/controllers/ProductCSVController.php
@@ -66,7 +66,8 @@ class ProductCSVController extends BaseController
     public function insertData($destinationPath,$files)
     {
         $productCSVRepo = App::make('ProductRepository');        
-        $excel = App::make('excel');           
+        $excel = App::make('excel');
+        $data = [];
         foreach($files as $file) {
 
             $product = $excel->selectSheets("Products")->load("./public/misc/$file"); 
@@ -78,13 +79,21 @@ class ProductCSVController extends BaseController
             $optionalAttributesObject = $attributes->get();
             $shipmentObject = $shipments->get();
             $imagesObject = $images->ignoreEmpty()->get();
-            $result = $this->ProductCSVService->insertData($productsObject, $optionalAttributesObject, $shipmentObject, $imagesObject);
-            
-            $data[]  = $result;
-            if (File::exists($destinationPath.$file)) {
-                File::delete($destinationPath.$file);
-            }                    
+
+            if(count($productsObject) > 0
+               && count($optionalAttributesObject) > 0
+               && count($shipmentObject) > 0
+               && count($imagesObject)) {
+               $data[] =  $this->ProductCSVService->insertData($productsObject, $optionalAttributesObject, $shipmentObject, $imagesObject);
+                if (File::exists($destinationPath.$file)) {
+                    File::delete($destinationPath.$file);
+                }
+            }
+            else {
+                continue;
+            }
         }
+        
         if(isset($data[0]["dataNotFound"])) {
             $this->ProductCSVService->removeErrorData($productsObject);
             return Response::json(['error' => $data]);
